assertion: exepcting 3.

+++ src/com/redhat/qe/katello/tests/api/Z_InitScriptLSBTest.java
@@ -62,7 +62,7 @@ public class Z_InitScriptLSBTest extends KatelloTestScript{
 		res = servertasks.execute_remote(String.format("service %s restart",KATELLO_SERVICENAME));
 		Assert.assertEquals(res.getExitCode(), new Integer(0),"Restarting of service (from stopped mode)");
 		res = servertasks.execute_remote(String.format("service %s status",KATELLO_SERVICENAME));
-		Assert.assertEquals(res.getExitCode(), new Integer(0),"Status command");
+		Assert.assertEquals(res.getExitCode(), new Integer(3),"Status command");
 		// restart service from the started mode
 		servertasks.execute_remote(String.format("service %s stop",KATELLO_SERVICENAME));
 		servertasks.execute_remote(String.format("export KATELLO_HOME=/usr/lib/katello; service %s start",KATELLO_SERVICENAME));
